+++ alembic.ini
@@ -61,7 +61,7 @@ version_path_separator = os  # Use os.pathsep. Default configuration used for ne
 # are written from script.py.mako
 # output_encoding = utf-8
 
-sqlalchemy.url = sqlite:///database.sqlite
+sqlalchemy.url = postgresql+pg8000://postgres:123@localhost/sts
 
 
 [post_write_hooks]
