+++ alembic/versions/5a65299b659c_add_enum.py
@@ -1,0 +1,30 @@
+"""add_enum
+
+Revision ID: 5a65299b659c
+Revises: 56286d60c53f
+Create Date: 2024-12-21 01:34:11.895469
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '5a65299b659c'
+down_revision: Union[str, None] = '56286d60c53f'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('request_stats', sa.Column('r_priority', sa.Enum('LOW', 'MEDIUM', 'HIGH', 'CRITICAL', name='requestpriorityenum'), nullable=False))
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('request_stats', 'r_priority')
+    # ### end Alembic commands ###
