+++ src/sts/stats/models.py
@@ -24,4 +24,5 @@ class RequestStat(BaseModel):
     hits: Mapped[int] = mapped_column(default=0)
     update_dt: Mapped[datetime.datetime] = mapped_column(default=func.now)
     errors: Mapped[int] = mapped_column(default=0)
-    priority: Mapped[RequestPriorityEnum] = mapped_column(sqlalchemy.Enum(RequestPriorityEnum, native_enum=False), nullable=False, default=RequestPriorityEnum.LOW)
+    r_priority: Mapped[RequestPriorityEnum] = mapped_column(sqlalchemy.Enum(RequestPriorityEnum),
+                                                          nullable=False, default=RequestPriorityEnum.LOW)
